+++ jobbergate-api/CHANGELOG.md
@@ -3,7 +3,7 @@
 This file keeps track of all notable changes to jobbergate-api
 
 ## Unreleased
-
+- Added clone capability to job submissions so they can be resubmitted to the cluster when needed [PENG-1676, ASP-4597]
 
 ## 5.4.0a2 -- 2024-11-06
 - Added back `libpq-dev` and `gcc` to the Dockerfile
@@ -18,7 +18,6 @@ This file keeps track of all notable changes to jobbergate-api
   - For Job Script files
   - For Job Script Template files
   - For Job Script Template workflow files
-
 
 ## 5.3.0 -- 2024-09-09